Update LICENSE
Shift to the MIT license. As the Unity asset hasn't been updated since `4.16` and apparently the project has been abandoned, it would be a nice move to release it into the public officially, so we can keep developing it and adjusting it to our needs. It has always been free, anyway. But as by the included license, it is governed by the asset store EULA which doesn't really make sense (anymore). 

btw: the "official" homepage https://master-toolkit.com isn't reachable anymore (and redirects to some very suspicious pages).

Signed-off-by: loptrinho <93672287+loptrinho@users.noreply.github.com>